// src/types/gitHubUser.ts
import * as z from "zod";

export const GitHubUserSchema = z.object({
  login: z.string(),
  id: z.number(),
  node_id: z.string(),
  avatar_url: z.string(),
  gravatar_id: z.string(),
  url: z.string(),
  html_url: z.string(),
  followers_url: z.string(),
  following_url: z.string(),
  gists_url: z.string(),
  starred_url: z.string(),
  subscriptions_url: z.string(),
  organizations_url: z.string(),
  repos_url: z.string(),
  events_url: z.string(),
  received_events_url: z.string(),
  type: z.string(),
  user_view_type: z.string(),
  site_admin: z.boolean(),
  name: z.string(),
  company: z.null(),
  blog: z.string(),
  location: z.string(),
  email: z.null(),
  hireable: z.null(),
  bio: z.string(),
  twitter_username: z.null(),
  public_repos: z.number(),
  public_gists: z.number(),
  followers: z.number(),
  following: z.number(),
  created_at: z.coerce.date(),
  updated_at: z.coerce.date(),
});
export type GitHubUser = z.infer<typeof GitHubUserSchema>;